+++ dlt/destinations/impl/dremio/configuration.py
@@ -4,7 +4,7 @@ from typing import Final, Optional, Any, Dict, ClassVar, List
 from dlt.common.configuration import configspec
 from dlt.common.configuration.specs import ConnectionStringCredentials
 from dlt.common.destination.reference import DestinationClientDwhWithStagingConfiguration
-from dlt.common.libs.sql_alchemy import URL
+from dlt.common.libs.sql_alchemy_shims import URL
 from dlt.common.typing import TSecretStrValue
 from dlt.common.utils import digest128
