+++ dlt/destinations/impl/mssql/configuration.py
@@ -1,6 +1,6 @@
 import dataclasses
 from typing import Final, ClassVar, Any, List, Dict
-from dlt.common.libs.sql_alchemy import URL
+from dlt.common.libs.sql_alchemy_shims import URL
 
 from dlt.common.configuration import configspec
 from dlt.common.configuration.specs import ConnectionStringCredentials
